+++ lib/services/secure_storage/secure_storage_service.dart
@@ -1,14 +1,26 @@
-import 'package:mobileprism/services/secure_storage/secure_storage_provider.dart';
+import 'dart:convert';
+
+import 'package:flutter_secure_storage/flutter_secure_storage.dart';
+
+class StorageIsNotOpen implements Exception {}
+
+class ItemNotFoundInStorage implements Exception {}
 
 class SecureStorageService {
-  final SecureStorageProvider _provider;
+  final FlutterSecureStorage _storage = const FlutterSecureStorage();
 
-  SecureStorageService(this._provider);
+  SecureStorageService();
 
-  void openStorage() => _provider.openStorage();
+  Future<void> storeData(String key, Object value) async {
+    await _storage.write(key: key, value: jsonEncode(value));
+  }
 
-  Future<void> storeData(String key, Object value) =>
-      _provider.storeData(key, value);
-
-  Future<T> readData<T>(String key) => _provider.readData(key);
+  Future<T> readData<T>(String key) async {
+    if (await _storage.containsKey(key: key)) {
+      final String item = await _storage.read(key: key) ?? "";
+      return jsonDecode(item) as T;
+    } else {
+      throw ItemNotFoundInStorage();
+    }
+  }
 }
