+++ lib/views/login_view.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:mobileprism/constants/routes.dart';
+import 'package:mobileprism/services/secure_storage/secure_storage_service.dart';
 
 class LoginView extends StatefulWidget {
   const LoginView({Key? key}) : super(key: key);
@@ -9,6 +10,7 @@ class LoginView extends StatefulWidget {
 }
 
 class _LoginViewState extends State<LoginView> {
+  final SecureStorageService _secureStorageService = SecureStorageService();
   late final TextEditingController _hostnameController;
   late final TextEditingController _usernameController;
   late final TextEditingController _passwordController;
@@ -27,6 +29,12 @@ class _LoginViewState extends State<LoginView> {
     _usernameController.dispose();
     _passwordController.dispose();
     super.dispose();
+  }
+
+  Future<void> _storeCredentials() async {
+    await _secureStorageService.storeData("hostname", _hostnameController.text);
+    await _secureStorageService.storeData("username", _usernameController.text);
+    await _secureStorageService.storeData("password", _passwordController.text);
   }
 
   @override
@@ -87,7 +95,9 @@ class _LoginViewState extends State<LoginView> {
                     children: [
                       ElevatedButton(
                         key: const Key("loginbutton"),
-                        onPressed: () {
+                        onPressed: () async {
+                          await _storeCredentials();
+                          if (!mounted) return;
                           Navigator.pushReplacementNamed(
                             context,
                             homeRoute,
